+++ app/api/prog_image/entities/errors/unprocessable_entity.rb
@@ -1,0 +1,8 @@
+module ProgImage
+  module Entities
+    module Errors
+      class UnprocessableEntity < Base
+      end
+    end
+  end
+end