+++ app/api/prog_image/entities/image_key.rb
@@ -1,0 +1,7 @@
+module ProgImage
+  module Entities
+    class ImageKey < Grape::Entity
+      expose :key, documentation: { type: 'String', desc: 'Unique identifier for your image' }
+    end
+  end
+end