+++ app/api/prog_image/helpers/api_response_helper.rb
@@ -1,7 +1,22 @@
 module ProgImage
-  module ApiResponseHelper
-    def say_ok
-      status :no_content
+  module Helpers
+    module ApiResponseHelper
+      def say_ok
+        status :no_content
+      end
+
+      def say_not_found
+        status :not_found
+      end
+
+      def say_unprocessable_entity
+        status :unprocessable_entity
+      end
+
+      def say_unprocessable_entity_for(form)
+        say_unprocessable_entity
+        present_form_errors form, Entities::Errors::UnprocessableEntity
+      end
     end
   end
 end