+++ app/forms/prog_image/base_form.rb
@@ -1,0 +1,8 @@
+require 'active_model'
+
+module ProgImage
+  class BaseForm
+    include ::ActiveModel::Model
+    include Virtus.model
+  end
+end